Update Package structure
- remove traits
- remove PHP5 support
- update package architecture to ease adding v9.0 public API
