+++ .php_cs
@@ -33,5 +33,5 @@ return Symfony\CS\Config\Config::create()
     ->finder(
         Symfony\CS\Finder\DefaultFinder::create()
             ->in(__DIR__.'/src')
-            ->in(__DIR__.'/test')
+            ->in(__DIR__.'/tests')
     );
