+++ composer.json
@@ -18,11 +18,11 @@
         "issues": "https://github.com/thephpleague/csv/issues"
     },
     "require": {
-        "php" : ">=5.5.0",
+        "php" : ">=7.0",
         "ext-mbstring" : "*"
     },
     "require-dev": {
-        "phpunit/phpunit" : "^4.0",
+        "phpunit/phpunit" : "^5.2",
         "friendsofphp/php-cs-fixer": "^1.9"
     },
     "autoload": {
@@ -32,8 +32,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "League\\Csv\\Test\\": "test",
-            "lib\\": "examples\\lib"
+            "LeagueTest\\Csv\\": "tests"
         }
     },
     "scripts": {
@@ -43,7 +42,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "8.2-dev"
+            "dev-master": "9.x-dev"
         }
     }
 }
