--- examples/example.css
@@ -1,53 +1,0 @@
-html {
-    font:normal 1em/1.5 sans-serif;
-}
-/* taken and adapted from lea verou excellent blog theme */
-pre {
-    margin-bottom: 1em;
-    overflow: auto;
-    padding: .5em .8em;
-    border: 1px solid #aaa;
-    background-color: #e1e1e1;
-    color: #444;
-    font-size: 1.2em;
-    line-height: 1.5;
-    text-align: left;
-    -webkit-tab-size: 4;
-       -moz-tab-size: 4;
-        -ms-tab-size: 4;
-         -o-tab-size: 4;
-            tab-size: 4;
-}
-body {
-     margin:0 auto;
-     width:50%
-}
-.table-csv-data {
-    width:98%;
-    margin:0 auto 1em;
-    border-collapse:collapse;
-    padding:0;
-    font:normal 1em/1.5 sans-serif;
-}
-.table-csv-data td {
-    border-bottom:1px solid #ccc;
-    padding:.3em;
-    color:#222;
-    vertical-align: top;
-    text-align:left;
-}
-
-.table-csv-data tr:nth-child(even) td {
-    background-color:rgba(192, 192, 192, .2);
-}
-
-.table-csv-data tr:hover td {
-    background-color:#d4e8fc;
-}
-
-.with-header tbody tr:first-of-type td {
-    font-weight:bold;
-}
-.with-header tbody tr:first-of-type:hover td {
-    background-color:inherit;
-}
