+++ phpunit.xml
@@ -13,7 +13,7 @@
 
     <testsuites>
         <testsuite name="League CSV Test Suite">
-            <directory>test</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
 
@@ -24,7 +24,6 @@
     </filter>
 
     <logging>
-        <log type="tap" target="build/report.tap"/>
         <log type="junit" target="build/report.junit.xml"/>
         <log type="coverage-html" target="build/coverage" charset="UTF-8" yui="true" highlight="true"/>
         <log type="coverage-text" target="build/coverage.txt"/>
