+++ src/MapIterator.php
@@ -4,13 +4,15 @@
 *
 * @license http://opensource.org/licenses/MIT
 * @link https://github.com/thephpleague/csv/
-* @version 8.2.0
+* @version 9.0.0
 * @package League.csv
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */
-namespace League\Csv\Modifier;
+declare(strict_types=1);
+
+namespace League\Csv;
 
 use Iterator;
 use IteratorIterator;
@@ -51,6 +53,6 @@ class MapIterator extends IteratorIterator
     {
         $iterator = $this->getInnerIterator();
 
-        return call_user_func($this->callable, $iterator->current(), $iterator->key(), $iterator);
+        return ($this->callable)($iterator->current(), $iterator->key(), $iterator);
     }
 }
