+++ src/Plugin/ForbiddenNullValuesValidator.php
@@ -4,12 +4,14 @@
 *
 * @license http://opensource.org/licenses/MIT
 * @link https://github.com/thephpleague/csv/
-* @version 8.2.0
+* @version 9.0.0
 * @package League.csv
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */
+declare(strict_types=1);
+
 namespace League\Csv\Plugin;
 
 /**
@@ -28,7 +30,7 @@ class ForbiddenNullValuesValidator
      *
      * @return bool
      */
-    public function __invoke(array $row)
+    public function __invoke(array $row): bool
     {
         $res = array_filter($row, function ($value) {
             return null === $value;
