+++ src/StreamIterator.php
@@ -4,13 +4,15 @@
 *
 * @license http://opensource.org/licenses/MIT
 * @link https://github.com/thephpleague/csv/
-* @version 8.2.0
+* @version 9.0.0
 * @package League.csv
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */
-namespace League\Csv\Modifier;
+declare(strict_types=1);
+
+namespace League\Csv;
 
 use InvalidArgumentException;
 use Iterator;
@@ -106,7 +108,7 @@ class StreamIterator implements Iterator
      * @param string $enclosure
      * @param string $escape
      */
-    public function setCsvControl($delimiter = ',', $enclosure = '"', $escape = '\\')
+    public function setCsvControl(string $delimiter = ',', string $enclosure = '"', string $escape = '\\')
     {
         $this->delimiter = $this->filterControl($delimiter, 'delimiter');
         $this->enclosure = $this->filterControl($enclosure, 'enclosure');
@@ -123,7 +125,7 @@ class StreamIterator implements Iterator
      *
      * @return string
      */
-    private function filterControl($char, $type)
+    protected function filterControl(string $char, string $type)
     {
         if (1 == strlen($char)) {
             return $char;
@@ -139,12 +141,8 @@ class StreamIterator implements Iterator
      *
      * @param int $flags
      */
-    public function setFlags($flags)
+    public function setFlags(int $flags)
     {
-        if (false === filter_var($flags, FILTER_VALIDATE_INT, ['options' => ['min_range' => 0]])) {
-            throw new InvalidArgumentException('The flags must be a positive integer');
-        }
-
         $this->flags = $flags;
     }
 
@@ -160,7 +158,7 @@ class StreamIterator implements Iterator
      *
      * @return int
      */
-    public function fputcsv(array $fields, $delimiter = null, $enclosure = null, $escape = null)
+    public function fputcsv(array $fields, string $delimiter = null, string $enclosure = null, string $escape = null)
     {
         return fputcsv(
             $this->stream,
